+++ proteusPy/disulfide_tree.py
@@ -16,6 +16,9 @@ Part of the program proteusPy, https://github.com/suchanek/proteusPy,
 a Python package for the manipulation and analysis of macromolecules.
 
 Author: Eric G. Suchanek, PhD
+Affiliation: Flux-Frontiers, https://github.com/Flux-Frontiers
+License: BSD
+Last revised: 2026-03-23 -egs-
 """
 
 from __future__ import annotations